renamed src/user_interaction to src/vacancy_interaction
