+++ src/vacancy_interaction.py
